feat: fix docker for not mounting client db

+++ docker-entrypoint.sh
@@ -1,10 +1,12 @@
 #!/usr/bin/env bash
 set -e
 
-until mysqladmin ping -h"$DB_HOST" -P"$DB_PORT" --silent; do
-  echo "Waiting for DB at $DB_HOST:$DB_PORT..."
-  sleep 3
-done
+if [ "$SERVICE_PATH" != "web-client" ] && [ -n "$DB_HOST" ]; then
+  until mysqladmin ping -h"$DB_HOST" -P"$DB_PORT" --silent; do
+    echo "Waiting for DB at $DB_HOST:$DB_PORT..."
+    sleep 3
+  done
+fi
 
 cd /var/www
 
@@ -26,10 +28,12 @@ if [ "$SERVICE_PATH" = "auth-service" ]; then
    php artisan jwt:generate-certs --force
 fi
 
-until php artisan migrate:fresh --seed --force; do
-  echo "Waiting for DB at $DB_HOST:$DB_PORT..."
-  sleep 3
-done
+if [ "$SERVICE_PATH" != "web-client" ] && [ -n "$DB_HOST" ]; then
+  until php artisan migrate:fresh --seed --force; do
+    echo "Waiting for DB at $DB_HOST:$DB_PORT..."
+    sleep 3
+  done
+fi
 
 if [ "$SERVICE_PATH" = "web-client" ] && [ -f package.json ]; then
   echo "→ Installing JS dependencies for web-client…"
